Slight oops.  Roof patterns are identified by the PngInstance in the Common Block of the catalog resource.  Everywhere else, I check whether PngInstance is set and use it - except I missed it for replacing the PNG!
Also took the opportunity to:
a) replace lots of
((AApiVersionedFields)selectedItem.Resource["CommonBlock"].Value)["someVal"]
with
selectedItem.Resource["CommonBlock.someVal"]
which makes the code easier to read (and saves a call to the library - not that it's slow).
b) get rid of array catalogTypes (which I was subscripting into to get type numbers) and use an enum instead, again making the code much more readable.  This lead to some rework on the menu code.