Fix prompts involving "current Source" and "current Target" languages to say what they are.
Source-level: btnRevert and btnRevertAll have been renamed to btnCopyToTarget and btnCopyToAll to avoid confusion.
Source-level: the innards of copying from source to target or all has been rewritten.

Clicking copy to Target explains what the source and target languages are.
Clicking copy to All explains what the source language is.
If the source language is actually missing, the prompt to say "Use xxx instead?" tells you what the source langage was.
You should now only get prompted once for any of that lot.

When switching languages (and in several other situations), if the source language doesn't exist, the prompt to use an alternative also tells you the source language.  It only bothers prompting if there is an alternative, too.
When switching languages (and in several other situations), if the target language doesn't exist, the prompt to create it tells you what the target language is.